+++ src/JSSoft.Commands/AssemblyInfo.cs
@@ -1,0 +1,13 @@
+// <copyright file="AssemblyInfo.cs" company="JSSoft">
+//   Copyright (c) 2024 Jeesu Choi. All Rights Reserved.
+//   Licensed under the MIT License. See LICENSE.md in the project root for license information.
+// </copyright>
+
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("JSSoft.Commands.Tests, PublicKey=002400000480000094000000060200000" +
+                              "0240000525341310004000001000100b95aa034365ba94cb2597445ba4a19b4cbe" +
+                              "54fe4402c3a7c99211bb321127c9c75d7a01523e1770a28886d857f5d9f1623d1d" +
+                              "5ceb772bfabaa1b2037cc470b087d66dd5120758034f8eb84a352c577316a522dd" +
+                              "95fd01bd868040176a3fd8d15ce3e3fbec56e5f833c5ccd1eddfa008f16714501b" +
+                              "fba78ad788095faedcb7dba")]