+++ src/JSSoft.Commands/CommandMethodDescriptor.cs
@@ -6,35 +6,67 @@
 using System.Diagnostics;
 using System.Threading;
 using System.Threading.Tasks;
+using JSSoft.Commands.Exceptions;
+using JSSoft.Commands.Extensions;
 
 namespace JSSoft.Commands;
 
-public abstract class CommandMethodDescriptor(MethodInfo methodInfo)
+public sealed class CommandMethodDescriptor
 {
-    public abstract string Name { get; }
+    private readonly PropertyInfo? _validationPropertyInfo;
+    private readonly MethodInfo? _completionMethodInfo;
 
-    public abstract string[] Aliases { get; }
+    public CommandMethodDescriptor(MethodInfo methodInfo)
+    {
+        if (methodInfo.DeclaringType is null)
+        {
+            throw new CommandDeclaringTypeNullException(methodInfo);
+        }
 
-    public abstract string DisplayName { get; }
+        MethodInfo = methodInfo;
+        Usage = CommandDescriptor.GetUsage(methodInfo);
+        MethodName = methodInfo.Name;
+        IsAsync = methodInfo.IsAsync();
+        Name = methodInfo.GetName();
+        Aliases = methodInfo.GetAliases();
+        DisplayName = methodInfo.GetDisplayName();
+        Members = CommandDescriptor.GetMemberDescriptorsByMethodInfo(methodInfo);
+        Category = AttributeUtility.GetCategory(methodInfo);
+        _validationPropertyInfo = methodInfo.GetValidationPropertyInfo();
+        _completionMethodInfo = methodInfo.GetCompletionMethodInfo();
+    }
 
-    public abstract string MethodName { get; }
+    public string Name { get; }
 
-    public abstract bool IsAsync { get; }
+    public string[] Aliases { get; }
 
-    public abstract CommandMemberDescriptorCollection Members { get; }
+    public string DisplayName { get; }
 
-    public abstract string Category { get; }
+    public string MethodName { get; }
 
-    public MethodInfo MethodInfo { get; } = methodInfo;
+    public bool IsAsync { get; }
 
-    public CommandUsage Usage { get; } = CommandDescriptor.GetUsage(methodInfo);
+    public CommandMemberDescriptorCollection Members { get; }
+
+    public string Category { get; }
+
+    public MethodInfo MethodInfo { get; }
+
+    public CommandUsage Usage { get; }
 
     internal bool CanExecute(object instance) => OnCanExecute(instance);
 
     internal object? Invoke(object instance, CommandMethodInstance methodInstance)
     {
         var parameters = methodInstance.GetParameters();
-        return OnInvoke(instance, parameters);
+        if (MethodInfo.DeclaringType!.IsAbstract && MethodInfo.DeclaringType.IsSealed is true)
+        {
+            return MethodInfo.Invoke(null, parameters);
+        }
+        else
+        {
+            return MethodInfo.Invoke(instance, parameters);
+        }
     }
 
     internal Task InvokeAsync(
@@ -44,7 +76,6 @@ public abstract class CommandMethodDescriptor(MethodInfo methodInfo)
         IProgress<ProgressInfo> progress)
     {
         var parameters = methodInstance.GetParameters(cancellationToken, progress);
-
         if (OnInvoke(instance, parameters) is Task task)
         {
             return task;
@@ -53,20 +84,74 @@ public abstract class CommandMethodDescriptor(MethodInfo methodInfo)
         throw new UnreachableException();
     }
 
-    internal string[] GetCompletionInternal(
-        object instance, CommandMemberDescriptor memberDescriptor, string find)
+    internal string[] GetCompletionsInternal(
+        object instance, CommandMemberDescriptor memberDescriptor)
     {
-        if (memberDescriptor.GetCompletionInternal(instance, find) is string[] items)
+        if (memberDescriptor.GetCompletionsInternal(instance) is string[] items)
         {
             return items;
         }
 
-        return GetCompletion(instance, [memberDescriptor, find]);
+        return GetCompletions(instance, [memberDescriptor]);
     }
 
-    protected abstract object? OnInvoke(object instance, object?[] parameters);
+    private object? OnInvoke(object instance, object?[] parameters)
+    {
+        if (MethodInfo.DeclaringType!.IsAbstract && MethodInfo.DeclaringType.IsSealed is true)
+        {
+            return MethodInfo.Invoke(null, parameters);
+        }
+        else
+        {
+            return MethodInfo.Invoke(instance, parameters);
+        }
+    }
 
-    protected virtual bool OnCanExecute(object instance) => true;
+    private bool OnCanExecute(object instance)
+    {
+        if (_validationPropertyInfo?.GetValue(instance) is bool @bool)
+        {
+            return @bool;
+        }
 
-    protected virtual string[] GetCompletion(object instance, object?[] parameters) => [];
+        return true;
+    }
+
+    private string[] GetCompletions(object instance, object?[] parameters)
+    {
+        if (_completionMethodInfo is not null)
+        {
+            return InvokeCompletionMethod(instance, parameters);
+        }
+
+        return [];
+    }
+
+    private string[] InvokeCompletionMethod(object instance, object?[] parameters)
+    {
+        try
+        {
+            var value = _completionMethodInfo!.Invoke(instance, parameters);
+            if (value is string[] items)
+            {
+                return items;
+            }
+            else if (value is Task<string[]> task)
+            {
+                if (task.Wait(CommandSettings.AsyncTimeout) is false)
+                {
+                    return [];
+                }
+
+                return task.Result;
+            }
+
+            throw new UnreachableException();
+        }
+        catch (Exception e)
+        {
+            Trace.TraceError($"{e}");
+            return [];
+        }
+    }
 }